Forbid connection to sqlite3 in memory db without explicit flag
A pool connection to in memory database would create a different database
per connection, data is not shared between these. There is no guarentee
which connection will be used for any given .query() hence what data is
available is not deterministic.

Throw error when connecting under these conditions (but added flag to
force this behaviour - maybe could be useful for some weird cache???)