Remove ConnectionPool, module now directly export ConnectionPromise
Idea is that you create a queryable (pool, connection, transaction) directly with any-db. You then wrap it with a ConnectionPromise and use as you would have before this change. This means that people already using any-db can continue as they are, and just wrap the queryable with an AnyDbQ ConnectionPromise where it makes sense to do so.